For editing of AccessImportConfig.xml file
Fixing the defect related to LDAP exception.

Fixed the issue of null in method write.

Removed the hard coding of Active Directory attributes to be fetched.

added the map for multiple data sources

changed the code for correct flow

Made the code changes as per  the new xml and bean.