added code to handle exceptions in method writeXML
Added the functionality to take care of error when the user import file is not present in "IN" folder.

Added code to take into account the fields added to Users table in db.

Code changes for added/modified columns to Users Table.

added code for moving file HRData.csv

Changed the code so that the" updatedate" of only the users that are updated, gets updated in table "Users".

Made changes to the code so that terminationpolicy is read correctly