added code to handle exceptions in method writeXML
Added the functionality to take care of error when the user import file is not present in "IN" folder.

Added code to take into account the fields added to Users table in db.

Code changes for added/modified columns to Users Table.

added code for moving file HRData.csv

Changed the code so that the" updatedate" of only the users that are updated, gets updated in table "Users".

Made changes to the code so that terminationpolicy is read correctly

made entry in table ResourceImportErrors for any import errors

Changed the code after the alteration is made in positions of fields in UserImportConfigTemplate.xml to ensure the correct mapping of fields to the corresponding columns in HRData.csv

changed the code so that the statusdesciption is populated correctly

Added tag <Status> inside <StatusField> tag so that terminatepolicy object is properly initiated.

Removed errors caused due to not taking the latest code